조회 테스트 및 오타 수정
