+++ src/main/java/com/gaaji/block/controller/BlockRetriveController.java
@@ -1,29 +1,33 @@
 package com.gaaji.block.controller;
 
+import java.util.List;
+
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.gaaji.block.applicationservice.BlockDeleteService;
+import com.gaaji.block.applicationservice.BlockRetriveService;
 import com.gaaji.block.controller.dto.BlockedUserRequest;
+import com.gaaji.block.domain.Block;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @RestController
-public class BlockDeleteControllar {
+public class BlockRetriveController {
 
-	private final BlockDeleteService blockDeleteService;
+	private final BlockRetriveService blockRetriveService;
 
 
-    @DeleteMapping("/block")
-    public ResponseEntity<Void> createBlock(@RequestHeader(HttpHeaders.AUTHORIZATION) String userId,
-            @RequestBody BlockedUserRequest dto){
-    	blockDeleteService.deleteBlock(userId, dto);
-        return ResponseEntity.status(HttpStatus.OK).build();
+    @GetMapping("/block")
+    public ResponseEntity<List<Block>> retriveBlock(@RequestHeader(HttpHeaders.AUTHORIZATION) String userId){
+
+    	return ResponseEntity.ok(blockRetriveService.retirveBlock(userId));
     }
 }
