+++ src/test/java/com/gaaji/block/applicationservice/BlockRetriveServiceTest.java
@@ -2,18 +2,19 @@ package com.gaaji.block.applicationservice;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.List;
+
+import org.junit.jupiter.api.Test;
+
 import com.gaaji.block.adaptor.AuthServiceClient;
 import com.gaaji.block.controller.dto.BlockedUserRequest;
 import com.gaaji.block.domain.Block;
 import com.gaaji.block.domain.UserId;
 import com.gaaji.block.repository.BlockRepository;
-import java.util.List;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-class BlockCreateServiceTest {
+public class BlockRetriveServiceTest {
 
-    @Test
+	@Test
     void 정상_생성() throws Exception{
         BlockRepository blockRepository = new FakeBlockRepository();
         AuthServiceClient authServiceClient = new StubAuthServiceClient();
@@ -21,9 +22,12 @@ class BlockCreateServiceTest {
 
         blockCreateService.createBlock("userId", new BlockedUserRequest("blockedUserId"));
 
-        List<Block> blocks= blockRepository.findAllByUserId(UserId.of("userId"));
+        
+        BlockRetriveService blockRetriveService = new BlockRetriveService(blockRepository);
+        List<Block> blocks= blockRetriveService.retirveBlock("userId");
 
+        
+        
         assertThat(blocks.size()).isSameAs(1);
     }
-
 }
